requirements_steamlit.txt
altair==6.0.0
attrs==26.1.0
blinker==1.9.0
branca==0.8.2
cachetools==7.0.5
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
filelock==3.25.2
folium==0.20.0
fonttools==4.62.1
fsspec==2026.3.0
gitdb==4.0.12
GitPython==3.1.46
idna==3.11
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
narwhals==2.18.1
networkx==3.6.1
numpy==2.4.3
opencv-python==4.13.0.92
packaging==26.0
pandas==2.3.3
pi_heif==1.3.0
pillow==12.1.1
plotly==6.6.0
polars==1.39.3
polars-runtime-32==1.39.3
protobuf==6.33.6
psutil==7.2.2
pyarrow==23.0.1
pydeck==0.9.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
requests==2.33.0
rpds-py==0.30.0
scipy==1.17.1
six==1.17.0
smmap==5.0.3
streamlit==1.55.0
streamlit-folium==0.27.1
sympy==1.14.0
tenacity==9.1.4
toml==0.10.2
torch==2.11.0
torchvision==0.26.0
tornado==6.5.5
typing_extensions==4.15.0
tzdata==2025.3
ultralytics==8.4.30
ultralytics-thop==2.0.18
urllib3==2.6.3
watchdog==6.0.0
xyzservices==2025.11.0
